Made changes to the naming structure of companants in navigation and logo. Added in brand logo to nav bar
