--- src/components/Logo/Logo.jsx
@@ -1,11 +1,0 @@
-
-
-const Logo = () => {
-  return (
-    <div className="logo">
-        Logo
-    </div>
-  )
-}
-
-export default Logo;
