+++ src/components/Logo/index.jsx
@@ -1,0 +1,5 @@
+const Logo = () => {
+  return <div className="logo">Logo</div>;
+};
+
+export default Logo;
