--- src/components/Navigation/Navigation.jsx
@@ -1,11 +1,0 @@
-import Logo from '../Logo/Logo';
-
-const Navigation = () => {
-  return (
-    <>
-     {/* replace whole content with boostrap navbar */}
-    </>
-  )
-}
-
-export default Navigation;
