+++ src/components/Navigation/_navigation.scss
@@ -1,4 +1,12 @@
 nav {
+  // margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  background-color: #eebe87;
+  margin-bottom: 20px;
+}
 
-  
+.logo {
+  width: 200px;
+  height: auto;
 }
