+++ src/components/Navigation/index.jsx
@@ -1,0 +1,12 @@
+import logo from "../../assets/logo.png";
+
+const Navigation = () => {
+  return (
+    <nav>
+      {/* replace whole content with boostrap navbar */}
+      <img src={logo} className="logo" />
+    </nav>
+  );
+};
+
+export default Navigation;
