+++ src/components/index.jsx
@@ -1,17 +1,9 @@
-import Navigation from './Navigation/Navigation';
-import Logo from "./Logo/Logo";
-import Aside from './Aside/Aside';
+import Navigation from "./Navigation";
+import Logo from "./Logo";
+import Aside from "./Aside/Aside";
 import Search from "./Search/Search";
-import MainContent from './MainContent/MainContent';
-import Favourites from './Favourites/Favourites';
-import Likes from './Likes/Likes';
+import MainContent from "./MainContent/MainContent";
+import Favourites from "./Favourites/Favourites";
+import Likes from "./Likes/Likes";
 
-export {
-  Navigation,
-  Logo,
-  Aside,
-  Search,
-  MainContent,
-  Favourites,
-  Likes,
-}
+export { Navigation, Logo, Aside, Search, MainContent, Favourites, Likes };
